estudiante/estudiante_app/serializers.py
from rest_framework import serializers
from .models import Estudiante

class EstudianteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Estudiante
        fields = ['rut', 'nombre_completo', 'edad', 'curso']