estudiante/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.1.1
requests==2.31.0